Cleanup and rename SklearnToDMLMapper.
